Package documentation

HierarchicalTemporalMemory.𝕊𝕢range

HierarchicalTemporalMemory.CellActivity

HierarchicalTemporalMemory.Hypercube

HierarchicalTemporalMemory.Hypersphere

HierarchicalTemporalMemory.InhibitionRadius

HierarchicalTemporalMemory.LengthfulIter

HierarchicalTemporalMemory.ProximalSynapses

HierarchicalTemporalMemory.SPParams

HierarchicalTemporalMemory.Sequence

HierarchicalTemporalMemory.SpatialPooler

HierarchicalTemporalMemory.SpatialPooler

HierarchicalTemporalMemory.TMParams

HierarchicalTemporalMemory.TemporalMemory

HierarchicalTemporalMemory.Truesof

HierarchicalTemporalMemory.𝕊𝕢

HierarchicalTemporalMemory.bitarray

HierarchicalTemporalMemory.hcat!!

HierarchicalTemporalMemory.predict!

HierarchicalTemporalMemory.reverse_simpleArithmetic

HierarchicalTemporalMemory.sp_activate

HierarchicalTemporalMemory.sparse_foreach

HierarchicalTemporalMemory.step!

HierarchicalTemporalMemory.step!

HierarchicalTemporalMemory.step_boost!

HierarchicalTemporalMemory.vcat!!

HierarchicalTemporalMemory.@percolumn

HierarchicalTemporalMemory.@percolumn

HierarchicalTemporalMemory.Hypercube — Type

Hypercube iterates over a hypercube of radius γ around xᶜ, bounded inside the space {1..sz}ᴺ.

Examples

julia> hc= Hypercube((3,3),1,(10,10));

julia> foreach(println, hc)
(2, 2)
(3, 2)
(4, 2)
(2, 3)
(3, 3)
(4, 3)
(2, 4)
(3, 4)
(4, 4)

HierarchicalTemporalMemory.Hypersphere — Type

Hypersphere is approximated with a Hypercube for simplicity

See also: Hypercube

HierarchicalTemporalMemory.SPParams — Type

SPParams{Nin,Nsp} holds the algorithm parameters for a spatial pooler with nomenclature similar to source

The dimension parameters are problem-specific and should be the first to be specified.

The tuning parameters have sensible defaults, which should be . All gated features are enabled by default.

Parameters

Dimensions

szᵢₙ, szₛₚ: input/output dimensions

γ: receptive field radius (how large an input area an output minicolumn maps to)

Algorithm tuning

s: average output sparsity

prob_synapse: probability for each element of the szᵢₙ × szₛₚ space to be a synapse. Elements that roll below this value don't form a synapse and don't get a permanence value. If this is very low, the proximal synapses matrix can become sparse.

θ_permanence01: synapse permanence connection threshold

p⁺_01,p⁻_01: synapse permanence adaptation rate (see ProximalSynapses)

θ_stimulus_activate: minicolumn absolute activation threshold

Tboost: boosting mechanism's moving average filter period

β: boosting strength

Feature gates

enable_local_inhibit

enable_learning

enable_boosting

HierarchicalTemporalMemory.SpatialPooler — Type

SpatialPooler{Nin,Nsp} is a learning algorithm that decorrelates the features of an input space, producing a Sparse Distributed Representation (SDR) of the input space. If defines the proximal connections of an HTM layer.

Examples

sp= SpatialPooler(SPParams(szᵢₙ=(600,), szₛₚ=(2048,)))
z= Random.bitrand(600)  # input
activation= sp(z)
# or, to adapt the SP to the input as well:
activation= step!(sp,z)

Properties

It's called SpatialPooler because input patterns that share a large number of co-active neurons (i.e., that are spatially similar) are grouped together into a common output representation. It is designed to achieve a set of computational properties that support further downstream computations with SDRs, including:

preserving topology of the input space by mapping similar inputs to similar outputs

continuously adapting to changing statistics of the input stream

forming fixed sparsity representations

being robust to noise

being fault tolerant

Source

Algorithm overview

Mapping I/O spaces

The spatial pooler maps an input space x to an output space y of minicolumns through a matrix of proximal synapses. The input space can optionally have a topology, which the spatial pooler will preserve by mapping output minicolumn yᵢ to a subset of the input space, a Hypercube around center xᶜ. julia xᶜ(yᵢ)= floor.(Int, (yᵢ.-1) .* (szᵢₙ./szₛₚ)) .+1`

Output activation

Calculate the overlap o(yᵢ) by propagating the input activations through the proximal synapses and adjusting by boosting factors b (control mechanism that spreads out the activation pattern across understimulated neurons)

Inhibition Z between yᵢ (local/global): competition where only the top-K yᵢ with the highest o(yᵢ) win; ensures sparsity

Activate winning yᵢ > activation threshold (θ_stimulus_activate)

See also: sp_activate

State variables

synapses: includes the synapse permanence matrix Dₚ

åₜ: [boosting] moving (in time) average (in time) activation of each minicolumn

åₙ: [boosting] moving (in time) average (in neighborhood) activation of each minicolumn

φ: the adaptible radius of local inhibition

See also: ProximalSynapses, TemporalMemory

HierarchicalTemporalMemory.SpatialPooler — Method

Return the activation pattern of the SpatialPooler for the given input activation.

z: CellActivity

Example

sp= SpatialPooler(SPParams(szᵢₙ=(600,), szₛₚ=(2048,)))
z= Random.bitrand(600)  # input
activation= sp(z)
# or, to adapt the SP to the input as well:
activation= step!(sp,z)

For details see: sp_activate

HierarchicalTemporalMemory.TMParams — Type

TMParams holds the algorithm parameters for a Temporal Memory with nomenclature similar to source

HierarchicalTemporalMemory.TemporalMemory — Type

TemporalMemory learns to predict sequences of input Sparse Distributed Representations (SDRs), usually generated by a SpatialPooler. It learns to represent each input symbol in the temporal context of the symbols that come before it in the sequence, using the individual neurons of each minicolumn and their distal synapses.

When considering a neuron layer with proximal and distal synapses, the spatial pooler is a way to activate and learn the proximal synapses, while the temporal memory is a way to activate and learn the distal synapses.

TM activation

An SDR input activates some minicolumns of the neuron layer.

See also: TMState

HierarchicalTemporalMemory.predict! — Method

predict!(classifier::SDRClassifier, Π,target::Int; enable_learning=true)=

Predict probability that the output represents each bucket of an arithmetic encoder. 0<=predict!<=1

HierarchicalTemporalMemory.reverse_simpleArithmetic — Method

reverse_simpleArithmetic(bucketProbDist, algorithm,params)

Reverses the simple arithmetic encoder, given the same parameters. Inputs a probability distribution across all buckets and collapses it to a single arithmetic value, representing the most likely estimation in the timeseries' domain (like a defuzzifier)

Arguments

bucketProbDist [nbucket]: discrete probability of each bucket [0-1]

algorithm: {'mean','mode','highmean'} 'highmean' is the mean of the highest-estimated values

HierarchicalTemporalMemory.step! — Method

step!(s::ProximalSynapses, z,a, params) adapts the proximal synapses' permanences with a hebbian learning rule on input z and activation a. The adaptation has a causal and an anticausal component:

If the postsynaptic neuron fires and the presynaptic fired too, the synapse is strengthened

If the postsynaptic neuron fires, but the presynaptic didn't, the synapse is weakened

See alse: ProximalSynapses

HierarchicalTemporalMemory.step! — Method

step!(sp::SpatialPooler, z::CellActivity) evolves the Spatial Pooler to the next timestep by evolving each of its constituents (synapses, boosting, inhibition radius) and returns the output activation.

See also: sp_activate

HierarchicalTemporalMemory.𝕊𝕢range — Constant

𝕊𝕢range is the domain of connection permanence quantization

HierarchicalTemporalMemory.CellActivity — Type

Type of neuron layer activations

HierarchicalTemporalMemory.InhibitionRadius — Type

The inhibition radius of a Spatial Pooler's columns is a dynamical system that evolves under the influence of other elements of the Spatial Pooler. It provides an init (constructor) and a step! function.

HierarchicalTemporalMemory.LengthfulIter — Type

Iterator transformations, like filters, don't keep the length info. LengthfulIter is a wrapper that allows length info known programmatically to be used.

HierarchicalTemporalMemory.ProximalSynapses — Type

ProximalSynapses{SynapseT<:AnySynapses,ConnectedT<:AnyConnection} are the feedforward connections of 2 neuron layers.

Used in the context of the SpatialPooler.

Description

The neurons of both layers are expected to form minicolumns which share the same feedforward connections. The synapses are binary: they don't have a scalar weight, but either conduct (1) or not (0). Instead, they have a permanence value Dₚ ∈ (0,1] and a connection threshold θ.

Initialization

Let presynaptic (input) neuron xᵢ and postsynaptic (output) neuron yᵢ, and a topological I/O mapping xᵢ(yᵢ) := Hypercube(yᵢ). ∀

Synapse adaptation

They adapt with a hebbian learning rule. The adaptation has a causal and an anticausal component:

If the postsynaptic neuron fires and the presynaptic fired too, the synapse is strengthened

If the postsynaptic neuron fires, but the presynaptic didn't, the synapse is weakened

The synaptic permanences are clipped at the boundaries of 0 and 1.

A simple implementation of the learning rule would look like this, where z: input, a: output

learn!(Dₚ,z,a)= begin
  Dₚ[z,a]  .= (Dₚ[z,a].>0) .* (Dₚ[z,a]   .⊕ p⁺)
  Dₚ[.!z,a].= (Dₚ[z,a].>0) .* (Dₚ[.!z,a] .⊖ p⁻)
end

Type parameters

They allow a dense or sparse matrix representation of the synapses

SynapseT: DenseSynapses or SparseSynapses

ConnectedT: DenseConnection or SparseConnection

See also: DistalSynapses, SpatialPooler, TemporalMemory

HierarchicalTemporalMemory.Sequence — Type

Sequence(;δ,init) is an easy way to define sequences with transition function δ and starting condition init as an iterator.

Examples

Fibonacci sequence:

fibonacci_δ(a,b)= b, a+b;
fibonacci_init= 0,1;
Lazy.@as _1 HTM.Sequence(fibonacci_δ, fibonacci_init) IterTools.take(_1, 13) foreach(println, _1)

# output
0
1
1
2
3
5
8
13
21
34
55
89
144

Factorial sequence:

factorial_δ(a,b)= a*b, b+1;
factorial_init= 1,1;
Lazy.@as _1 HTM.Sequence(factorial_δ, factorial_init) IterTools.take(_1, 13) foreach(println, _1)

# output
1
1
2
6
24
120
720
5040
40320
362880
3628800
39916800
479001600

HierarchicalTemporalMemory.Truesof — Type

Iterate over the trues of a BitArray

Examples

julia> b= Random.bitrand(5)
5-element BitArray{1}:
 1
 0
 0
 1
 1

julia> foreach(i-> print(string(i)*" "), HTM.Truesof(b))
1 4 5

HierarchicalTemporalMemory.𝕊𝕢 — Type

𝕊𝕢 is the type of connection permanences, defining their quantization domain

HierarchicalTemporalMemory.bitarray — Method

bitarray(dims, idx)

Create a bitarray with true only at idx.

HierarchicalTemporalMemory.hcat!! — Method

hcat!!(s::SparseMatrixCSC, I,V)

WARNING: s is invalid after this operation!!! Its symbol should be reassigned to the output of hcat!!. Append new columns to s with 1 value V[c] at row I[c] each. Because SparseMatrixCSC is immutable, return a new, valid SparseMatrixCSC that points to the original's data structures. For large matrices, this is much faster than hcat.

HierarchicalTemporalMemory.sp_activate — Method

sp_activate(sp::SpatialPooler{Nin,Nsp}, z) calculates the SP's output activation for given input activation z.

Algorithm

Overlap o(yᵢ)

Propagate the input activations through the proximal synapses (matrix multiply)

Apply boosting factors b: control mechanism that spreads out the activation pattern across understimulated neurons (homeostatic excitability control)

Inhibition Z between yᵢ (local/global): competition where only the top-k yᵢ with the highest o(yᵢ) win; ensures sparsity The competition happens within an area around each neuron.

k: number of winners depends on desired sparsity (s, see SPParams) and area size

θ_inhibit: inhibition threshold per neighborhood $:= o(k-th y)$

Z(o(yᵢ)): convolution of o with θ_inhibit

Activate winning yᵢ > activation threshold (θ_stimulus_activate)

Info

The local sparsity s tends to the sparsity of the entire layer as it grows larger, but for small values or small inhibition radius it diverges, because of the limited & integral number of neurons winning in each neighborhood. This could be addressed by tie breaking, but it doesn't seem to have much practical importance.

HierarchicalTemporalMemory.sparse_foreach — Method

sparse_foreach(f, s::SparseMatrixCSC,columnIdx)

Iterate SparseMatrix s and apply f columnwise (f(s,nzrange,rowvals))

HierarchicalTemporalMemory.step_boost! — Method

step_boost!(sp::SpatialPooler,a) evolves the boosting factors b (see sp_activate). They depend on:

åₜ: moving average in time activation of each minicolumn

åₙ: moving average in neighborhood activation of each minicolum

HierarchicalTemporalMemory.vcat!! — Method

vcat!!(s::SparseMatrixCSC, J,V)

WARNING: s is invalid after this operation!!! Its symbol should be reassigned to the output of vcat!!. Append new rows to s with 1 value V[r] at column J[r] each. Because SparseMatrixCSC is immutable, return a new, valid SparseMatrixCSC that points to the original's data structures. For large matrices, this is much faster than vcat.

HierarchicalTemporalMemory.@percolumn — Macro

@percolumn(f,a,b,k,Ncol)

Macro to apply f elementwise and concatenate the results.

a: vector of size [Ncol*k], column-major

b: vectors of size Ncol

HierarchicalTemporalMemory.@percolumn — Macro

@percolumn(reduce,a,k,Ncol)

Macro to reduce a per column.